The write vocabulary moves to the format floor
A pure refactor, and the third time this exact move has been made: `@olai/ops`'
`request.ts` becomes `@olai/format`'s `writing.ts`, and the ops layer keeps a
re-export under its own names. `committing.ts` did it for git, `reading.ts` and
`searching.ts` did it for the four reads (#167, #170); this is the other half of
the ops vocabulary, and it is here for the same structural reason:
`@olai/surface` may not import `@olai/ops` and `@olai/ops` may not import
`@olai/surface`, so a vocabulary they must agree about has exactly one home that
is neither. `ops.run` is about to be a surface procedure, so its request and its
answer are now values that cross that floor.

TWO RENAMES AT THE MOVE, both saying what the thing is against the neighbours it
now has:

  - `Request` → `WriteRequest`. Among `NodeRequest`, `SubtreeRequest`,
    `SearchRequest` and `CommitRequest`, a bare `Request` is the one that does
    not say which question it is — and reads like their supertype, which it is
    not. `@olai/ops` re-exports it as `Request`, which is the right name in a
    layer that has exactly one kind.
  - `Applied` → `WriteResult`, and this one is the trap the `reads-on-the-floor`
    node recorded, met head on. `@olai/surface` exports an `Applied` of its own
    and it is a genuinely DIFFERENT type: the keyboard's answer, which adds
    `undo` and drops `summary`, `sort`, `captured` and `rev`. Two `Applied`s,
    both about a write that landed, both carrying `id`/`title`/`nudge` — the
    `Outline`/`OutlineSummary` collision again, and the nastier kind, because
    the shapes look compatible at a glance. Named for the family it actually
    joins (`CommitResult`, `PushResult`) it can never be mistaken for the other.

`WriteResult` and `Minted` become SCHEMAS rather than the interfaces they were,
which is the one thing here that is not a move: a procedure's answer has to be a
shape both ends decode against, not a type only the producing end can check.
Field for field identical, `optionalKey` where the interface said `?`.

The planner does not move, exactly as the matcher and the walks did not. What is
here is the shape; `NESTING` comes with it because it is a fact about the JSON
Schema an MCP host can read, and it lives beside the schema that unrolls it.

`just typecheck` and `just test` (1599 pass, 0 fail) green; no call site changed,
because none had to.
