+++ packages/format/README.md
@@ -352,6 +352,38 @@ suite, one `.test.ts` beside each module:
 bun test packages/format
 ```
 
+## What a write ASKS for, and what one that landed says
+
+`src/writing.ts` is the third time this move was made — after `committing.ts`
+and `reading.ts`/`searching.ts`, and for the identical reason. It holds the
+seventeen request schemas, the `WriteRequest` union over them, and
+`WriteResult`, which is what an op that succeeded answers with.
+
+It moved here from `@olai/ops` when the writes went onto the surface: `ops.run`
+is a procedure now, so the request an agent sends and the answer it gets back
+are values that CROSS this floor. The ops layer produces them, the surface
+carries them, and `@olai/ops` re-exports the two under its own names —
+`Request` and `Applied` — the way `Pending` and `CommitRequest` are already
+re-exported by their consumers rather than declared twice.
+
+Both are renamed at the move, and the second rename is the load-bearing one:
+
+- `Request` → **`WriteRequest`**, because among `NodeRequest`,
+  `SubtreeRequest`, `SearchRequest` and `CommitRequest` a bare `Request` is the
+  one that does not say which question it is;
+- `Applied` → **`WriteResult`**, because `@olai/surface` exports an `Applied` of
+  its own and it is a DIFFERENT type — the keyboard's answer, which adds `undo`
+  and drops `summary`, `sort`, `captured` and `rev`. Two things called
+  `Applied`, both about a write that landed, both carrying `id`/`title`/`nudge`,
+  is the `Outline`/`OutlineSummary` trap again: not a rename away from a compile
+  error, a rename away from a plausible one.
+
+The PLANNER stays in `@olai/ops`, exactly as the matcher and the walks did.
+What is here is only the shape, plus the one number the shape imposes —
+`NESTING`, the depth to which `add_node`'s `children` is unrolled, which lives
+beside the schema that unrolls it because it is a fact about the JSON Schema an
+MCP host can read rather than about the format.
+
 ## Writing is here too, and for the same reason reading is
 
 `src/write.ts` is the only thing in olai that turns records back into bytes,
